Refactoring code base - using context api - custom hook
