+++ src/components/Options.jsx
@@ -1,4 +1,8 @@
-function Options({ question, answer, dispatch }) {
+import { useQuiz } from "../contexts/QuizContext";
+
+function Options({ question }) {
+  const { answer, dispatch } = useQuiz();
+
   const hasAnswered = answer !== null;
   return (
     <div className="options">
