+++ src/components/Progress.jsx
@@ -1,4 +1,7 @@
-function Progress({ index, numQuestions, points, maxPossiblePoints, answer }) {
+import { useQuiz } from "../contexts/QuizContext";
+
+function Progress() {
+  const { index, numQuestions, points, maxPossiblePoints, answer } = useQuiz();
   return (
     <header className="progress">
       <progress value={index + Number(answer !== null)} max={numQuestions} />
